+++ src/main/java/com/project/velo/entity/Review.java
@@ -39,7 +39,7 @@ public class Review {
     @Min(value = 1, message = "Минимальная оценка — 1")
     @Max(value = 5, message = "Максимальная оценка — 5")
     @Column(nullable = false)
-    private BigDecimal score;
+    private Integer score;
 
     @Size(max = 1000, message = "Отзыв не может быть длиннее 1000 символов")
     @Column(columnDefinition = "TEXT")
